add OptimizedParquetFilterSuite
